@@ -10,44 +10,37 @@
         "@fortawesome/free-solid-svg-icons": "^6.1.1",
         "@fortawesome/react-fontawesome": "^0.1.18",
         "@reduxjs/toolkit": "^1.8.2",
-        "animate.css": "^4.1.1",
         "antd": "^4.21.3",
-        "bootstrap": "5.1.0",
-        "gh-pages": "^3.2.3",
-        "jquery": "3.6.0",
-        "react": "17.0.2",
-        "react-device-detect": "^1.17.0",
-        "react-dom": "17.0.2",
-        "react-helmet": "6.1.0",
-        "react-redux": "^7.2.4",
-        "react-router-dom": "5.2.1",
-        "react-scripts": "4.0.3",
-        "redux": "^4.1.1",
-        "redux-thunk": "^2.3.0"
+        "gh-pages": "^4.0.0",
+        "jquery": "^3.6.0",
+        "react": "^18.2.0",
+        "react-device-detect": "^2.2.2",
+        "react-dom": "^18.2.0",
+        "react-helmet": "^6.1.0",
+        "react-redux": "^8.0.2",
+        "react-router-dom": "^6.3.0",
+        "redux": "^4.2.0",
+        "redux-thunk": "^2.4.1"
     },
     "devDependencies": {
-        "@types/history": "4.7.9",
-        "@types/jquery": "^3.5.6",
-        "@types/node": "^16.7.10",
-        "@types/react": "17.0.18",
-        "@types/react-dom": "17.0.9",
-        "@types/react-google-recaptcha": "^2.1.2",
-        "@types/react-helmet": "^6.1.2",
-        "@types/react-redux": "^7.1.18",
-        "@types/react-router": "5.1.16",
-        "@types/react-router-dom": "5.1.8",
-        "@typescript-eslint/parser": "4.29.2",
+        "@types/history": "^5.0.0",
+        "@types/jquery": "^3.5.14",
+        "@types/node": "^18.0.0",
+        "@types/react": "^18.0.14",
+        "@types/react-dom": "^18.0.5",
+        "@types/react-helmet": "^6.1.5",
+        "@types/react-redux": "^7.1.24",
+        "@types/react-router-dom": "^5.3.3",
+        "@typescript-eslint/parser": "^5.28.0",
         "@vitejs/plugin-react": "^1.3.2",
-        "cross-env": "7.0.3",
-        "eslint": "7.32.0",
-        "eslint-config-react-app": "6.0.0",
-        "eslint-plugin-flowtype": "5.9.0",
-        "eslint-plugin-import": "2.24.0",
-        "eslint-plugin-jsx-a11y": "6.4.1",
-        "eslint-plugin-react": "7.24.0",
-        "nan": "2.15.0",
+        "eslint": "^8.18.0",
+        "eslint-config-react-app": "^7.0.1",
+        "eslint-plugin-flowtype": "^8.0.3",
+        "eslint-plugin-import": "^2.26.0",
+        "eslint-plugin-jsx-a11y": "^6.5.1",
+        "eslint-plugin-react": "^7.30.0",
         "tailwindcss": "^3.1.3",
-        "typescript": "4.3.5",
+        "typescript": "^4.7.4",
         "vite": "^2.9.12"
     },
     "eslintConfig": {
