@@ -1,38 +1,38 @@
-import { Redirect, Route, Router, Switch, useHistory } from 'react-router';
+import { Routes, Route, Navigate, } from 'react-router-dom';
+import { Provider } from 'react-redux';
+import { Store } from './Store';
+import { Fragment } from 'react';
+
 import Layout from './components/Layout';
 import { Home } from './pages/Home';
 import { Skills } from './pages/Skills';
 import { Activities } from './pages/Activities';
 import { Contact } from './pages/Contact';
 
-import "bootstrap/dist/css/bootstrap.min.css"
-
 import "antd/dist/antd.min.css"
+import "./assets/styles/Overrides.css"
 import "./assets/styles/Styles.Compiled.css"
 
-
-import { connect, Provider } from 'react-redux';
-import { AppDispatch, ChangeLanguage, RootState, Store } from './Store';
 
 export default function App() {
     const random = (Math.random()) * 2
     const lang = Math.floor(random) ? "En" : "Fa"
     return (
-        <Provider store={Store}>
-            <Layout >
-                <Router history={useHistory()}>
-                    <Switch>
-                        <Route path='/Resume/:lang/Home/' component={Home} />
-                        <Route path='/Resume/:lang/Skills/' component={Skills} />
-                        <Route path='/Resume/:lang/Activities/' component={Activities} />
-                        <Route path='/Resume/:lang/Contact/' component={Contact} />
+        <Fragment>
+            <Provider store={Store} >
+                <Layout>
+                    <Routes>
+                        <Route path='/Resume/:lang/Home/' element={<Home />} />
+                        <Route path='/Resume/:lang/Skills/' element={<Skills />} />
+                        <Route path='/Resume/:lang/Activities/' element={<Activities />} />
+                        <Route path='/Resume/:lang/Contact/' element={<Contact />} />
 
                         <Route>
-                            <Redirect to={`/Resume/${lang}/Home`} />
+                            <Navigate to={`/Resume/${lang}/Home`} />
                         </Route>
-                    </Switch>
-                </Router>
-            </Layout>
-        </Provider>
+                    </Routes>
+                </Layout>
+            </Provider>
+        </Fragment>
     );
 }
