@@ -1,13 +1,14 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import App from './App';
 
 const baseUrl = document.getElementsByTagName('base')[0].getAttribute('href') as string;
-const rootElement = document.getElementById('app');
+const rootElement = document.getElementById('app')!;
 
-ReactDOM.render(
+const { render } = createRoot(rootElement);
+
+render(
     <BrowserRouter basename={baseUrl}>
         <App />
-    </BrowserRouter>,
-    rootElement);
+    </BrowserRouter>
+);
